# src/main/java/com/exercise/inventorymanager/item/model/Item.java
package com.exercise.inventorymanager.item.model;

public class Item {
    private Integer id;
    private String name;
    private Integer quantity;
    private Double value;

    public void setId(Integer id) {
        this.id = id;
    }

    public Integer getId() {
        return id;
    }

    public void setName(String name) {
        this.name = name;
    }

    public String getName() {
        return name;
    }

    public void setQuantity(Integer quantity) {
        this.quantity = quantity;
    }

    public Integer getQuantity() {
        return quantity;
    }

    public void setValue(Double value) {
        this.value = value;
    }

    public Double getValue() {
        return value;
    }
}